# Vectorized pattern-match branch selection