--- a/unison-runtime/src/Unison/Runtime/Serialize.hs
+++ b/unison-runtime/src/Unison/Runtime/Serialize.hs
@@ -167,18 +167,18 @@ getEnumMap :: (MonadGet m) => (EnumKey k) => m k -> m v -> m (EnumMap k v)
 getEnumMap gk gv = mapFromList <$> getList (getPair gk gv)
 
 putSmallEnumMap ::
-  (MonadPut m, EnumKey k) =>
-  (k -> m ()) ->
+  (MonadPut m) =>
+  (Word64 -> m ()) ->
   (v -> m ()) ->
-  SmallEnumMap k v ->
+  SmallEnumMap v ->
   m ()
 putSmallEnumMap pk pv m = putFoldable (putPair pk pv) (EC.smallEnumMapToList m)
 
 getSmallEnumMap ::
-  (MonadGet m, EnumKey k) =>
-  m k ->
+  (MonadGet m) =>
+  m Word64 ->
   m v ->
-  m (SmallEnumMap k v)
+  m (SmallEnumMap v)
 getSmallEnumMap gk gv = EC.smallEnumMapFromList <$> getList (getPair gk gv)
 
 putEnumSet :: (MonadPut m) => (EnumKey k) => (k -> m ()) -> EnumSet k -> m ()