@ChrisPenner 2024-09-03
Unfortunately this doesn't appear to speed things up (in fact it's slower) so tabling for now.

`run '@pchiusano/misc-benchmarks/chris:.patternMatchTests'`

```
trunk:
Pattern match D1
1.887µs

Pattern match D2
2.47µs

Pattern match D4
3.733µs

Pattern match D16
10.9µs

Ability handler 1 case
5.158µs

Ability handler 2 cases
7.677µs

Ability handler 4 cases
7.752µs

Ability handler 16 cases
14.68µs

Pattern match on numbers (100 cases)
673ns

Pattern match on Text
6.005µs
```

```
this-branch:

Pattern match D1
1.868µs

Pattern match D2
2.457µs

Pattern match D4
3.66µs

Pattern match D16
10.897µs

Ability handler 1 case
5.165µs

Ability handler 2 cases
7.765µs

Ability handler 4 cases
7.831µs

Ability handler 16 cases
14.891µs

Pattern match on numbers (100 cases)
824ns

Pattern match on Text
6.123µs
```

JSON deserialization:

`run '@mitchellwrosen/mbta/@mitchellwrosen/branch-for-chris:.runTheBenchmark'`

```
trunk:
read bytes: 0.177ms
utf8 decode byes: 0.171ms
json decode text: 499.497ms
jsonapi parse json: 171.497ms
mbta parse jsonapi: 185.568ms
```

```
this-branch:
utf8 decode byes: 0.165ms
json decode text: 518.767ms
jsonapi parse json: 172.558ms
mbta parse jsonapi: 193.953ms
```
```